Add navigation between login and register pages.
